Amount + delete + update

+++ src/main/java/sk/stuba/fei/uim/oop/assignment3/ProductResponse.java
@@ -9,12 +9,13 @@ public class ProductResponse {
     private String description;
     private int amount;
     private String unit;
-    private int price;
+    private Double price;
 
     public ProductResponse(Product product){
         this.id = product.getId();
         this.name = product.getName();
         this.description = product.getDescription();
+
         this.amount = product.getAmount();
         this.unit = product.getUnit();
         this.price = product.getPrice();
